Create code1
This will create an instance of the Calendar class and display the current month and year. Additionally, it allows you to add notes to specific days and displays them when you call the displayNotes() method.
Please note that this is just an example, and it doesn't have any UI. You'll have to create the UI to display the calendar and interact with the created object and methods.